Home layout and data sent together
